This TTF parsing was not what I needed
The SVG files I'm generating PDFs from list the font-family as 
whatever-the-basename of the TTF file is, but this one will try (and 
fail) to divine the name from headers in the file, and then add goofy
spacing to it to ensure they won't match. 

Nope. Just going to use the basename, kthx. YMMV.